+++ Examples/SimpleExample/example_app.cpp
@@ -301,8 +301,8 @@ void ExampleApp::render()
         .left = 0,
         .top = 0,
         .front = 0,
-        .right = static_cast<long>(dim.x),
-        .bottom = static_cast<long>(dim.y),
+        .right = dim.x,
+        .bottom = dim.y,
         .back = 0,
     };
     m_context->set_viewports(&cmd_list, 1, &viewport);
